+++ RELEASE_NOTES.md
@@ -1,10 +1,10 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v2.2.3
+This is the offical release of vpinfe v2.2.4
 
 ### What's New
-None
+- add addtional opions to vpx_config.py page
 
 ### Fixes
 - fix stale cache after saving. #4